requirements.txt
attrs==22.1.0
aws-cdk-lib==2.40.0
boto3==1.24.66
botocore==1.27.66
cattrs==22.1.0
constructs==10.1.94
distlib==0.3.6
exceptiongroup==1.0.0rc9
filelock==3.8.0
fsspec==2022.8.2
jmespath==1.0.1
jsii==1.67.0
numpy==1.23.2
pandas==1.4.4
platformdirs==2.5.2
publication==0.0.3
python-dateutil==2.8.2
pytz==2022.2.1
s3fs==0.4.2
s3transfer==0.6.0
six==1.16.0
typeguard==2.13.3
typing_extensions==4.3.0
urllib3==1.26.12
virtualenv==20.16.4